Split benchmark report diagnostics and sections
